% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunk_list.R
\name{chunk_list}
\alias{chunk_list}
\title{Diviser une liste ou vecteur en morceaux (chunks)}
\usage{
chunk_list(vec, chunk_size)
}
\arguments{
\item{vec}{Un vecteur ou une liste à diviser.}

\item{chunk_size}{Un entier indiquant la taille de chaque chunk.}
}
\value{
Une liste contenant les chunks.
}
\description{
Cette fonction divise un vecteur ou une liste en sous-listes (chunks) de taille spécifiée.
}
\examples{
chunk_list(1:10, 3)

}
