% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_groups_by_factor_level.R
\name{compare_groups_by_factor_level}
\alias{compare_groups_by_factor_level}
\title{Compare deux groupes d'objets scientifiques selon un facteur donné}
\usage{
compare_groups_by_factor_level(
  session,
  group_dict,
  experiment_id,
  group1_id,
  group2_id,
  factor
)
}
\arguments{
\item{session}{Session RSession d’accessibilité aux données (via API).}

\item{group_dict}{Liste contenant les groupes sous forme de data.frames.}

\item{experiment_id}{ID ou nom de l’expérience concernée.}

\item{group1_id}{ID du premier groupe à comparer (doit exister dans `group_dict`).}

\item{group2_id}{ID du second groupe à comparer.}

\item{factor}{Le nom du facteur discriminant (ex: "genotype", "treatment").}
}
\value{
Un `data.frame` tidy contenant les données combinées prêtes à être exploitées.
}
\description{
Cette fonction récupère les données pour deux groupes définis,
prépare un tableau tidy, puis affiche les courbes de comparaison
(trajectoires individuelles et moyennes ± écart-type).
}
