% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_data_by_variable_to_csv.R
\name{export_data_by_variable_to_csv}
\alias{export_data_by_variable_to_csv}
\title{Export data by variable to CSV files}
\usage{
export_data_by_variable_to_csv(
  var_exp,
  data,
  csv_filepath = "/home/abidri/Bureau/test_R/csv_groups"
)
}
\arguments{
\item{var_exp}{DataFrame contenant la metadata des variables, avec au moins les colonnes `URI` et `Name`.}

\item{data}{DataFrame contenant les données scientifiques, avec au moins les colonnes `variable`, `target`, `value` et `date`.}

\item{csv_filepath}{Chemin du dossier où les fichiers CSV seront enregistrés. Par défaut : "/home/abidri/Bureau/test_R/csv_groups".}
}
\value{
Une liste de DataFrames, chaque élément correspondant aux données filtrées d'une variable.
}
\description{
Cette fonction exporte les données scientifiques groupées par variable dans des fichiers CSV séparés.
Chaque variable aura son propre fichier CSV nommé d'après la variable.
}
