% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_group_objects.R
\name{extract_group_objects}
\alias{extract_group_objects}
\title{Extraire un groupe d'objets scientifiques par ID}
\usage{
extract_group_objects(group_dict, group_id, output_filename = NULL)
}
\arguments{
\item{group_dict}{Une liste nommée contenant des objets `data.frame`, typiquement issue d’un split par groupe.}

\item{group_id}{Le nom du groupe à extraire (doit correspondre à un nom de `group_dict`).}

\item{output_filename}{(optionnel) Chemin vers un fichier `.csv` où sauvegarder les données du groupe extrait.
Si `NULL`, aucune sauvegarde n'est effectuée.}
}
\value{
Le `data.frame` correspondant au groupe extrait.
}
\description{
Cette fonction extrait un sous-ensemble de données (`data.frame`) à partir d'une liste (dictionnaire)
indexée par identifiants de groupe. Elle peut aussi sauvegarder le groupe extrait dans un fichier `.csv`.
}
