% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_chunk_data.R
\name{fetch_chunk_data}
\alias{fetch_chunk_data}
\title{Fetch data by chunk using GraphQL query}
\usage{
fetch_chunk_data(chunk, experience, session)
}
\arguments{
\item{chunk}{Un vecteur de chaînes de caractères représentant les URIs des objets scientifiques à récupérer.}

\item{experience}{Un vecteur de chaînes de caractères indiquant les expériences concernées par la requête.}

\item{session}{Un objet de connexion contenant les informations d'authentification et l'URL GraphQL.}
}
\value{
Un data.frame (ou tibble) contenant les données des objets scientifiques récupérées.
}
\description{
Interroge une API GraphQL pour récupérer les données des objets scientifiques correspondant à un chunk d'URIs.
}
\examples{
\dontrun{
session <- login("user", "password", "https://example.com", "https://example.com/graphql")
fetch_chunk_data(c("uri1", "uri2"), c("exp1"), session)
}

}
