% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_by_os_uri_variable.R
\name{get_data_by_os_uri_variable}
\alias{get_data_by_os_uri_variable}
\title{Retrieve data by OS URI and variables}
\usage{
get_data_by_os_uri_variable(session, experience, df_os, ls_var_exp = NULL)
}
\arguments{
\item{session}{Un objet de connexion contenant les informations d'authentification et l'URL GraphQL.}

\item{experience}{Un vecteur ou liste d'expériences à interroger.}

\item{df_os}{DataFrame contenant au moins une colonne 'uri' listant les objets scientifiques à interroger.}

\item{ls_var_exp}{Optionnel. DataFrame des variables par expérience. Si NULL ou vide, la fonction lsVarByExp sera appelée.}
}
\value{
Aucune valeur retournée explicitement, les données sont exportées via export_data_by_variable_to_csv().
}
\description{
Récupère les données scientifiques associées à une liste d'URIs d'objets scientifiques, en traitant les variables d'expérience.
}
