% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNamesFromUri.R
\name{getNamesFromUri}
\alias{getNamesFromUri}
\title{Get Names from URI}
\usage{
getNamesFromUri(uri)
}
\arguments{
\item{uri}{Character string. The URI to look up.}
}
\value{
Character vector of name(s) associated with the URI.
}
\description{
Retrieves the name(s) associated with a given URI from the mapping.
}
