% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insertUri_Name.R
\name{insertUri_Name}
\alias{insertUri_Name}
\title{Insert URI-Name Mapping}
\usage{
insertUri_Name(uri, name)
}
\arguments{
\item{uri}{Character string. The URI to add.}

\item{name}{Character string. The name associated with the URI.}
}
\description{
Adds a new URI-name pair to the mapping if the URI doesn't already exist.
}
\examples{
\dontrun{
initUriName()
insertUri_Name("http://example.com/exp1", "Experiment 1")
}
}
