% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/login.R
\name{login}
\alias{login}
\title{Login to OpenSILEX}
\usage{
login(id, password, instance, urlGraphql, uri_csv = "uri_name.csv")
}
\arguments{
\item{id}{Character string. User identifier.}

\item{password}{Character string. User password.}

\item{instance}{Character string. Base URL of the OpenSILEX instance.}

\item{urlGraphql}{Character string. GraphQL endpoint URL.}

\item{uri_csv}{Character string. Path to URI-name mapping CSV file.}
}
\value{
An object of class "opensilex_connection" containing authentication
details and connection information.
}
\description{
Authenticates with an OpenSILEX instance and returns a connection object.
}
