% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsMoveByOs.R
\name{lsMoveByOs}
\alias{lsMoveByOs}
\title{Retrieve movement/position data of a scientific object within an experiment}
\usage{
lsMoveByOs(
  session,
  experiment_label,
  scientific_object_id,
  start_date = NULL,
  end_date = NULL,
  output_dir = NULL,
  verbose = TRUE
)
}
\arguments{
\item{session}{An opensilex_connection object.}

\item{experiment_label}{Character, experiment label.}

\item{scientific_object_id}{Character, ID of the scientific object.}

\item{start_date}{Character or POSIXct or NULL, start date filter (optional).}

\item{end_date}{Character or POSIXct or NULL, end date filter (optional).}

\item{output_dir}{Character or NULL, directory to save CSV file (optional).}

\item{verbose}{Logical, whether to print messages (default TRUE).}
}
\value{
A tibble data frame with movement records.
}
\description{
Retrieve movement/position data of a scientific object within an experiment
}
