% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsVarByFacility.R
\name{lsVarByFacility}
\alias{lsVarByFacility}
\title{Retrieve variables associated with a facility and optional date range}
\usage{
lsVarByFacility(
  session,
  facility_uri,
  date_beginning = NULL,
  date_end = NULL,
  csv_filepath = NULL
)
}
\arguments{
\item{session}{An opensilex_connection object.}

\item{facility_uri}{Character, URI of the facility.}

\item{date_beginning}{Character or NULL, start date filter (YYYY-MM-DD).}

\item{date_end}{Character or NULL, end date filter (YYYY-MM-DD).}

\item{csv_filepath}{Character or NULL, path to save CSV file.}
}
\value{
A tibble data frame with variable details.
}
\description{
Retrieve variables associated with a facility and optional date range
}
