% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data_for_comparison.R
\name{prepare_data_for_comparison}
\alias{prepare_data_for_comparison}
\title{Préparer les données tidy pour comparaison entre deux groupes}
\usage{
prepare_data_for_comparison(
  group_dict,
  session,
  experiment_name,
  group1_id,
  group2_id,
  factor
)
}
\arguments{
\item{group_dict}{Liste nommée de groupes (`data.frame`), généralement issue d’un split par facteur.}

\item{session}{Objet de session avec token et URL GraphQL pour les requêtes.}

\item{experiment_name}{Nom ou identifiant de l’expérience dans l’API.}

\item{group1_id}{Identifiant du premier groupe (nom dans `group_dict`).}

\item{group2_id}{Identifiant du second groupe.}

\item{factor}{Nom de la colonne utilisée pour définir les groupes (ex. `"genotype"`).}
}
\value{
Un `data.frame` tidy combinant les données des deux groupes, avec colonnes :
`Date`, `URI`, `Groupe`, `factor`, `Variable`, `Valeur`.
}
\description{
Cette fonction extrait les données de deux groupes d’objets scientifiques,
récupère leurs mesures, y ajoute les informations de groupe et de facteur,
puis retourne un `data.frame` en format long (`tidy`) prêt pour visualisation ou analyse.
}
