% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_trailing_nans_from_identifier.R
\name{remove_trailing_nans_from_identifier}
\alias{remove_trailing_nans_from_identifier}
\title{Nettoyer un identifiant de groupe}
\usage{
remove_trailing_nans_from_identifier(group_identifier)
}
\arguments{
\item{group_identifier}{Chaîne de caractères. Identifiant du groupe à nettoyer.}
}
\value{
Une chaîne de caractères correspondant à l’identifiant nettoyé. Retourne une chaîne vide si tout a été supprimé.
}
\description{
Supprime les suffixes \code{"_nan"} (non sensibles à la casse) à la fin d’un identifiant de groupe.
}
\examples{
remove_trailing_nans_from_identifier("WD_ZM1289_nan")  # "WD_ZM1289"
remove_trailing_nans_from_identifier("WW_123_nan_nan") # "WW_123"

}
