% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualise_group_data.R
\name{visualise_group_data}
\alias{visualise_group_data}
\title{Visualiser les données groupées par variable}
\usage{
visualise_group_data(df_variables, output_file = NULL)
}
\arguments{
\item{df_variables}{Une liste de `tibble`, chaque élément représentant une variable sous forme de `data.frame`
avec les colonnes `URI`, le nom de la variable, et `Date`.}

\item{output_file}{(Optionnel) Chemin vers un fichier image (ex: `"output.png"`). Si `NULL` ou vide, la visualisation n'est pas enregistrée.}
}
\value{
Aucun objet retourné. Affiche la figure combinée à l'écran et éventuellement sauvegarde le fichier.
}
\description{
Cette fonction crée une série de graphiques temporels (un par variable) pour les objets scientifiques,
avec des couleurs spécifiques par URI. Elle combine ensuite les graphiques en une seule figure à l'aide du package `patchwork`.
Si un chemin est fourni, la visualisation est enregistrée dans un fichier.
}
\details{
- Les graphiques affichent les données dans le temps pour chaque URI, colorées selon une palette cohérente.
- Le type de visualisation (points ou ligne) dépend de la densité et complétude des données.
- Les légendes sont regroupées et affichées en bas.
}
