% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_os.R
\name{get_data_os}
\alias{get_data_os}
\title{Récupère les données scientifiques pour une expérience donnée}
\usage{
get_data_os(session, exp_name, df_os)
}
\arguments{
\item{session}{Une liste contenant les informations de session, incluant l'URL GraphQL et le token d'authentification.}

\item{exp_name}{Nom ou identifiant de l'expérience.}

\item{df_os}{DataFrame contenant les objets scientifiques, avec au minimum une colonne `uri`.}
}
\value{
Un `tibble` combiné de toutes les données récupérées ou `NULL` en cas d'erreur.
}
\description{
Cette fonction utilise le plan `future::multisession` pour paralléliser la récupération
des données liées à une liste d'objets scientifiques (`df_os`) pour une expérience donnée (`exp_name`).
En cas d'erreur, elle capture proprement l'exception et renvoie `NULL`.
}
