% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insertUri_Name.R
\name{insertUri_Name}
\alias{insertUri_Name}
\title{Insert multiple URI-name pairs into the uri_name mapping.}
\usage{
insertUri_Name(uri_name_pairs)
}
\arguments{
\item{uri_name_pairs}{A DataFrame containing 'uri' and 'name' columns with the new URI-name pairs to be added.}
}
\description{
This function inserts multiple URI-name pairs at once into the uri_name mapping
(stored in `.pkg_env$uri_name`). It ensures that no duplicate URI-name pairs are inserted.
It also checks for consistency (URIs with multiple names or names with multiple URIs)
after insertion.
}
