% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/login.R
\name{login}
\alias{login}
\title{Login to OpenSILEX}
\usage{
login(id, password, instance, urlGraphql, uri_csv = "uri_name.csv")
}
\arguments{
\item{id}{User email}

\item{password}{Your password}

\item{instance}{Base instance URL (e.g., http://localhost:8080)}

\item{urlGraphql}{GraphQL URL endpoint}
}
\value{
A session token (or list) on success
}
\description{
Login to OpenSILEX
}
\examples{
# Exemple d'utilisation de la fonction login
session <- login(id = "ton_email@exemple.com",
                 password = "ton_mot_de_passe",
                 instance = "http://localhost:8080",
                 urlGraphql = "http://localhost:8080/graphql")
print(session$token)
}
