% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsFacilityByExp.R
\name{lsFacilityByExp}
\alias{lsFacilityByExp}
\title{Retrieve facilities used in experiments}
\usage{
lsFacilityByExp(
  session,
  exp_label = NULL,
  output_dir = NULL,
  verbose = TRUE,
  check_consistency = TRUE
)
}
\arguments{
\item{session}{An opensilex_connection object.}

\item{exp_label}{Character or NULL, experiment label filter (optional).}

\item{output_dir}{Character or NULL, directory to save CSV file (optional).}

\item{verbose}{Logical, whether to print messages (default TRUE).}

\item{check_consistency}{Logical, whether to check URI-Name consistency (default TRUE).}
}
\value{
A tibble data frame of facilities by experiment.
}
\description{
Retrieve facilities used in experiments
}
