% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_trailing_nans_from_identifier.R
\name{remove_trailing_nans_from_identifier}
\alias{remove_trailing_nans_from_identifier}
\title{Clean a Group Identifier}
\usage{
remove_trailing_nans_from_identifier(group_identifier)
}
\arguments{
\item{group_identifier}{Character string. Group identifier to be cleaned.}
}
\value{
A character string corresponding to the cleaned identifier.
  Returns an empty string if everything has been removed.
}
\description{
Removes trailing \code{"_nan"} suffixes (case-insensitive) at the end of a
group identifier.
}
\examples{
remove_trailing_nans_from_identifier("WD_ZM1289_nan")  # "WD_ZM1289"
remove_trailing_nans_from_identifier("WW_123_nan_nan") # "WW_123"

}
