% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UriCons.R
\name{UriCons}
\alias{UriCons}
\title{Check URI-Name Consistency}
\usage{
UriCons()
}
\description{
Checks for inconsistencies in the uri_name mapping:
- URIs associated with multiple names
- Names associated with multiple URIs
Displays warnings but does not modify the data.
}
